requirements.txt
boto3
numpy
keras
tensorflow==2.2
datetime
pandas
nltk
matplotlib
RPi.GPIO